Unify exception thrown in l3-agent-scheduler fails
Since you can only attach a single l3 agent to a router, when you try
to add another l3 agent to a router that already have one, the l3
agent scheduler raises an exception.

This fix removes the discrimination by id: either it is the same agent
or another one, the router can not be hosted and the same exception is
raised.

Change-Id: If832bbd4bf17e4e0c4720172aded4c9fffedc6fc
Fixes: bug #1154622